A few bug bugfixes around assumptions

+++ bash/new-home-ip.sh
@@ -49,18 +49,20 @@ if [[ "${ipv4}" != "${old4}" ]]; then
   sed -i "s/${old4}/${ipv4}/" ${temp}
   iptables-restore < ${temp}
   [[ $? -ne 0 ]] && echo "ERROR: iptables-restore failed" && exit 3
-  cat ${temp} > /etc/sysconfig/iptables
+  [[ -f /etc/sysconfig/iptables ]] && cat ${temp} > /etc/sysconfig/iptables
   rm -f ${temp}
 
   # TODO - Test transmission fix
   # Note that whitelisting ipv6 is not supported
   transConf='/var/lib/transmission/.config/transmission-daemon/settings.json'
-  grep -q ${old4} ${transConf}
-  if [[ $? -eq 0 ]]; then
-    systemctl stop transmission-daemon
-    sed -i "s/${old4}/${ipv4}/" ${transConf}
-    [[ $? -ne 0 ]] && echo "ERROR: transmission update failed" && exit 4
-    systemctl start transmission-daemon
+  if [[ -f "${transConf}" ]]; then
+    grep -q ${old4} ${transConf}
+    if [[ $? -eq 0 ]]; then
+      systemctl stop transmission-daemon
+      sed -i "s/${old4}/${ipv4}/" ${transConf}
+      [[ $? -ne 0 ]] && echo "ERROR: transmission update failed" && exit 4
+      systemctl start transmission-daemon
+    fi
   fi
 fi
 
@@ -78,7 +80,7 @@ if [[ "${ipv6}" != "${old6}" ]]; then
   sed -i "s/${old6}/${ipv6}/" ${temp}
   ip6tables-restore < ${temp}
   [[ $? -ne 0 ]] && echo "ERROR: iptables-restore failed" && exit 3
-  cat ${temp} > /etc/sysconfig/ip6tables
+  [[ -f /etc/sysconfig/ip6tables ]] && cat ${temp} > /etc/sysconfig/ip6tables
   rm -f ${temp}
 
 fi
